Fix missing cast.

+++ src/timeman.h
@@ -30,7 +30,7 @@
 class TimeManagement {
 public:
   void init(Search::LimitsType& limits, Color us, int ply);
-  void pv_instability(bool easy, int easyPlayed, double bestMoveChanges) { unstablePvFactor = 1.08 + bestMoveChanges - easy / (easyPlayed + 1); }
+  void pv_instability(bool easy, int easyPlayed, double bestMoveChanges) { unstablePvFactor = 1.08 + bestMoveChanges - double(easy) / (easyPlayed + 1); }
   int available() const { return int(optimumTime * unstablePvFactor * 0.72); }
   int maximum() const { return maximumTime; }
   int elapsed() const { return int(Search::Limits.npmsec ? Threads.nodes_searched() : now() - startTime); }
